refactor: split out more
