+++ src/sphinx_ts/directives/common.py
@@ -1,0 +1,10 @@
+"""Common Formatting Utilities for TypeScript Directives.
+
+Contains shared formatting methods for TypeScript methods and properties
+that can be used across different directive types.
+"""
+
+from __future__ import annotations
+
+# This module is reserved for future common functionality between directives.
+# Currently, all formatting logic is implemented directly in the base directive class.
